requirements.txt
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.15.0
cssselect2==0.5.0
defusedxml==0.7.1
numpy==1.22.3
opencv-python==4.5.5.62
Pillow==9.0.1
pkg_resources==0.0.0
pycparser==2.21
tinycss2==1.1.1
webencodings==0.5.1
